remove payload from readme.md
So this pattern of using payloads in an action are a bit of an anti-pattern that I have seen in some of the redux curriculum at learn and not one I would recommend using. I've never seen it at the two jobs I've worked at where we used Redux and React. It also seems counterintuitive, because an action is already a payload of information and we are adding unnecessary nesting inside of a JavaScript Object. Dan Abramov defines an action as, "Actions are payloads of information that send data from your application to your store." He also never uses the nested attribute of payload once in his documentation.